Up to 3.0: Output a PPM, vec3 Class defined
A simple nested for loop outputs specific coloured pixels to a PPM file. This is mainly used to check that the program is ready to work as intended.

The vec3 class has been defined, and utility functions have been appended to the file outside of the class body.

Notes have been made about the "inline" keyword.

Another thing to note is the use of "return *this;". The "this" keyword is dereferenced here because "this" is a pointer to an object. By dereferencing it, it returns the object itself.