Ray-Sphere Hit Detection
The program now detects whether a sphere is hit at any point along a ray. Sounds great! Well this includes behind the camera as well. There is also no shading or anything else fancy, it's a simple true or false, and if it hits the sphere then the pixel is coloured red instead of the background colour.