Simplified hit_sphere() code
Code within hit_sphere() has been simplified *MATHEMATICALLY*. As the b coefficient had a factor of -2, b was replaced with -2h, allowing for the quadratic formula to be simplified and the b coefficient (now named h) has the value dot(rayDirectionVector, centreToOriginVector), which only differs by not being multiplied by -2. This has affected the discriminant value, and the x value when solving the quadratic formula.

The a and c coefficients have also been rewritten, as a vector dotted against itself is equal to that vector's length squared. The values are still the same though.