Created interval class and changed files to use it
Interval class manages real-value intervals (min and max). Can return the range of an interval and whether a given value is within the interval (different functions for inclusive or exclusive of min and max).

Code within other files has been adjusted to use this interval class. For example, hittable::hit() previously included individual parameters of type double for the minimum and maximum values of a ray's t value (the subtract and add values from the quadratic equation). Now, these have been replaced by one parameter of type interval for the t values, constructed with the provided min and max.