Camera code moved to camera class
Much of the code in main was related to the camera and has been moved into a camera class. Some parts were also split into functions that did not already exist.

Main has been adjusted to remove the code that is now in the camera class, but a call to the camera's render function (passing in the hittable_list world) has been added.