Anti-Aliasing via pixel sampling
Each pixel now has multiple rays directed at random points within a square sample. The colour returned by each ray is added to a color instance, and the summed value for each pixel is divided by the pixel count to get the mean colour within that pixel's sample area.

The number of samples per pixel is defined in main().

A new interval::clamp() function has been created to ensure values can be kept within a specified range, such as colours in the range [0, 256), as seen in write_color() in color.h.

In order to generate random positions within a pixel sample area, pseudo-random numbers can be returned by random_double() in rtweekend.h. This function returns a real number in [0, 1) by default, or can be overflowed with min and max values to return a number in [min, max).